delegate registerCourses to course service
